app/requirements.txt
flask===3.0.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1